Moving getAvailableDatastores method to DatastoreDescriptors class.
